moved files + clangformat
