+++ test/include/dlaf_test/matrix/util_tile_blas.h
@@ -14,7 +14,7 @@
 
 #include "blas.hh"
 #include "dlaf/tile.h"
-#include "dlaf_test/util_tile.h"
+#include "dlaf_test/matrix/util_tile.h"
 #include "dlaf_test/util_types.h"
 
 namespace dlaf {
@@ -37,7 +37,7 @@ void set(Tile<T, Device::CPU>& tile, Func el, blas::Op op) {
       break;
 
     case blas::Op::Trans: {
-      auto op_el = [el](TileElementIndex i){ 
+      auto op_el = [el](TileElementIndex i) {
         i.transpose();
         return el(i);
       };
@@ -46,13 +46,13 @@ void set(Tile<T, Device::CPU>& tile, Func el, blas::Op op) {
     }
 
     case blas::Op::ConjTrans: {
-      auto op_el = [el](TileElementIndex i){ 
+      auto op_el = [el](TileElementIndex i) {
         i.transpose();
         return TypeUtilities<T>::conj(el(i));
       };
       set(tile, op_el);
       break;
-  }
+    }
   }
 }
 }
