+++ test/unit/communication/test_broadcast_matrix.cpp
@@ -13,7 +13,7 @@
 #include <gtest/gtest.h>
 
 #include "dlaf_test/helper_communicators.h"
-#include "dlaf_test/util_tile.h"
+#include "dlaf_test/matrix/util_tile.h"
 #include "dlaf_test/util_types.h"
 
 #include "dlaf/matrix.h"
