+++ test/unit/mc/test_cholesky.cpp
@@ -13,7 +13,7 @@
 #include "dlaf/communication/communicator_grid.h"
 #include "dlaf/matrix.h"
 #include "dlaf_test/comm_grids/grids_6_ranks.h"
-#include "dlaf_test/util_matrix.h"
+#include "dlaf_test/matrix/util_matrix.h"
 #include "dlaf_test/util_types.h"
 
 using namespace dlaf;
