+++ test/unit/test_blas_tile/test_gemm.h
@@ -17,8 +17,8 @@
 #include "dlaf/memory/memory_view.h"
 #include "dlaf/tile.h"
 #include "dlaf/util_blas.h"
-#include "dlaf_test/util_tile.h"
-#include "dlaf_test/util_tile_blas.h"
+#include "dlaf_test/matrix/util_tile.h"
+#include "dlaf_test/matrix/util_tile_blas.h"
 #include "dlaf_test/util_types.h"
 
 using namespace dlaf;
