+++ test/unit/test_matrix.cpp
@@ -14,7 +14,7 @@
 #include "gtest/gtest.h"
 #include "dlaf/communication/communicator_grid.h"
 #include "dlaf_test/comm_grids/grids_6_ranks.h"
-#include "dlaf_test/util_matrix.h"
+#include "dlaf_test/matrix/util_matrix.h"
 #include "dlaf_test/util_types.h"
 
 using namespace dlaf;
